+++ tests/test_device.py
@@ -14,6 +14,8 @@
 """
 Unit tests for Lightning devices creation.
 """
+import sys
+
 import numpy as np
 import pennylane as qml
 import pytest
@@ -66,8 +68,12 @@ def test_create_device_with_unsupported_mpi_buf_size():
 
 
 @pytest.mark.skipif(
-    device_name == "lightning.tensor",
-    reason="lightning.tensor doesn't support 0 wires.",
+    (device_name == "lightning.kokkos" and sys.platform == "win32"),
+    reason="lightning.kokkos doesn't support 0 wires on Windows.",
+)
+@pytest.mark.skipif(
+    device_name in ["lightning.gpu", "lightning.tensor"],
+    reason=device_name + " doesn't support 0 wires.",
 )
 def test_device_init_zero_qubit():
     """Test the device initialization with zero-qubit."""